+++ src/core/middleware/api.js
@@ -2,7 +2,8 @@ import { Schema, arrayOf, normalize } from 'normalizr';
 // import { camelizeKeys } from 'humps';
 import 'isomorphic-fetch';
 
-const API_ROOT = 'https://addons-dev.allizom.org/api/v3/';
+// const API_ROOT = 'https://addons-dev.allizom.org/api/v3/';
+const API_ROOT = 'http://olympia.dev/api/v3/';
 
 // Fetches an API response and normalizes the result JSON according to schema.
 // This makes every API response have the same shape, regardless of how nested it was.
