+++ src/search/containers/CurrentSearchPage.js
@@ -1,19 +1,20 @@
 import React, { PropTypes } from 'react';
 import { connect } from 'react-redux';
-import { loadSearch } from 'search/actions';
 
-import SearchForm from '../components/SearchForm';
-import SearchResults from '../components/SearchResults';
+import SearchForm from 'search/components/SearchForm';
+import SearchResults from 'search/components/SearchResults';
+import { loadSearch } from 'search/actions';
 import { gettext as _ } from 'core/utils';
 
 import 'search/css/SearchPage.scss';
 
 function loadData(props) {
+  props.loadSearch(props.query);
 }
 
 class SearchPage extends React.Component {
   static propTypes = {
-    handleSearch: PropTypes.func.isRequired,
+    loadSearch: PropTypes.func.isRequired,
     loading: PropTypes.bool.isRequired,
     results: PropTypes.arrayOf(PropTypes.shape({
       slug: PropTypes.string.isRequired,
@@ -22,13 +23,17 @@ class SearchPage extends React.Component {
     query: PropTypes.string,
   }
 
+  // componentWillMount() {
+  //   loadData(this.props);
+  // }
+
   render() {
-    const { handleSearch, loading, results, query } = this.props;
+    const { loading, results, query } = this.props;
 
     return (
       <div className="search-page">
         <h1>{_('Add-on Search')}</h1>
-        <SearchForm onSearch={handleSearch} query={query} />
+        <SearchForm onSearch={this.props.loadSearch} query={query} />
         <SearchResults results={results} query={query} loading={loading} />
       </div>
     );
@@ -39,21 +44,18 @@ export function mapStateToProps(state, context) {
   return Object.assign({}, state.search, { query });
 }
 
-export function mapDispatchToProps(dispatch, context) {
-  return {
-    handleSearch: (query) => {
-      context.history.push({
-        pathname: '/search',
-        query: {q: query},
-      });
-      dispatch(loadSearch({ query }));
-    },
+function handleSearch(query) {
+  return (dispatch) => {
+    context.history.push({
+      pathname: '/search',
+      query: {q: query},
+    });
+    dispatch(loadSearch({ query }));
   };
 }
 
-const CurrentSearchPage = connect(
-  mapStateToProps,
-  mapDispatchToProps,
-)(SearchPage);
+const CurrentSearchPage = connect(mapStateToProps, {
+  loadSearch,
+})(SearchPage);
 
 export default CurrentSearchPage;
